Initial AI Refactor
