+++ Assets/Scripts/Knockback.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec28e5d8cddadb34181fab740f4447f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
