+++ Assets/Scripts/PlayerAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 673c16ed6c14c964392ff63b0900bc79
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
